pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sp-lakehouse-backup"
version = "0.1.0"
description = "Multi-server PostgreSQL backup/restore orchestration using LangChain + LangGraph with Ollama LLMs and MCP REST APIs"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "SP Lakehouse Backup Team"}
]
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-ollama>=0.3.0",
    "pydantic>=2.0.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]

[project.scripts]
sp-backup = "cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*", "venv*", "backups*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
